Use Integer/String IDs and relax ticket fields
Convert ticket-related identifiers and history IDs from Long to Integer/String and relax several ticket/user fields to strings or nullable columns. Update controllers and services to accept the new path variable types and handle Authentication principals (improved principal resolution in NotificationController), and add authentication to ticket status updates. TicketService now nulls resourceId to avoid FK violations, adds DataIntegrityViolation handling, and ensures changedByUserId falls back to the resolved user. Add CRITICAL priority, expose optional resource/location inputs in the frontend, surface avatar URL handling in AdminDashboard, and include a sample avatar asset.

+++ backend/src/main/java/backend/repository/TicketStatusHistoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface TicketStatusHistoryRepository extends JpaRepository<TicketStatusHistory, Integer> {
-    List<TicketStatusHistory> findByTicketTicketIdOrderByChangedAtAsc(Long ticketId);
+    List<TicketStatusHistory> findByTicketTicketIdOrderByChangedAtAsc(Integer ticketId);
 }
